Respect WWW_SCHEME configuration
My Cabot server requires SSL connections and responds with a 502 error if incoming requests do not comply. Before this change, the plugin would correctly ring phones, but no message was spoken upon picking up. Also, the Twilio web console sent an alert saying that they got a 502 when trying to retrieve the twiml to play back the voice.

Test Plan:
I manually modified the models.py file in the `site-packages/cabot_alert_twilio` directory to include this change.
Then, the next twilio API calls made had the correct callback URL with the `https` scheme.
The Twilio web console showed that the request and response was as expected.
Answering the phone call resulted in the correct message being spoken by Twilio's text-to-voice.